double precision for bspline
git-svn-id: svn://222.29.86.177/wanghan/projects/gpu-md/branches/trials@796 a28d8dd4-3dfb-4ec0-b98a-e5cf011a4b41

+++ spme/src/SPMERec.cu
@@ -318,9 +318,9 @@ calQMat (const IntVectorType K,
 	value.y = uu.y - meshIdx.y;
 	value.z = uu.z - meshIdx.z;
 	for (IndexType jj = 0; jj < order; ++jj){
-	  Mnx[jj] = BSplineValue (order, value.x);
-	  Mny[jj] = BSplineValue (order, value.y);
-	  Mnz[jj] = BSplineValue (order, value.z);
+	  Mnx[jj] = BSplineValue (order, double(value.x));
+	  Mny[jj] = BSplineValue (order, double(value.y));
+	  Mnz[jj] = BSplineValue (order, double(value.z));
 	  value.x += 1.;
 	  value.y += 1.;
 	  value.z += 1.;
